backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
websockets==14.1
sqlalchemy==2.0.36
alembic==1.14.0
asyncpg==0.30.0
redis==5.2.1
pydantic==2.10.3
pydantic-settings==2.7.0
httpx==0.28.1
jinja2==3.1.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
